Updated README. Reverted cleanDB.php (didnt understood that the cl.php file was just used as a lock). Accordingly, renamed cl.php to db.lck to comply with Unix standards of using .lck files for files used as a lock.

+++ cleanDB.php
@@ -4,33 +4,34 @@
 
 	$verbose=$config["verbose"];
 
-#$filename = 'cl.php';
-#$s = 'DO NOT REMOVE THIS FILE! Unless you want to clean your DB!';
+	$filename = 'db.lck';
+	$s = 'DO NOT REMOVE THIS FILE! Unless you want to clean your DB!';
 
-#if (file_exists($filename)){
-#	echo "Db is already cleaned!";
-#} else {
-	$db = new SQLite3($config["temperature_DB"]);
-	//DROP TABLE
-	$db->query('DROP TABLE rpi_temp');
-
-	if($verbose)
-		echo "DB Droped! <br>";
-	$db->query("CREATE TABLE 'rpi_temp' ('id' INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, 'temp' REAL, 'time' TEXT, 'cpu' REAL)");
-
-	if($verbose)
-		echo "DB created! <br>";
-	$db = NULL;
+	if (file_exists($filename)){
+		if($verbose) {
+			echo "DB cleaning is locked. Delete $filename to clean database!";
+		}
+	} 
+	else {
+		$db = new SQLite3($config["temperature_DB"]);
+		//DROP TABLE
+		$db->query('DROP TABLE rpi_temp');
 	
-	require 'measuretemp.php';
-
-	if($verbose)
-		echo "First values inserted! Good luck:) <br>";
+		if($verbose)
+			echo "DB Droped! <br>";
+		$db->query("CREATE TABLE 'rpi_temp' ('id' INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, 'temp' REAL, 'time' TEXT, 'cpu' REAL)");
 	
-	#file_put_contents($filename, $s);
+		if($verbose)
+			echo "DB created! <br>";
+		$db = NULL;
+		
+		require 'measuretemp.php';
 	
-#}
-
-
+		if($verbose)
+			echo "First values inserted! Good luck:) <br>";
+		
+		file_put_contents($filename, $s);
+	
+	}
 ?>
